requirements.txt
streamlit
pandas
numpy
plotly
sentence-transformers
torch
Pillow
pyyaml
streamlit-authenticator
bcrypt